requirements.txt
requests==2.32.3
python-dotenv==1.0.1
supabase==2.10.0
postgrest==0.17.0
tenacity==8.5.0
tqdm==4.66.4
